+++ spec/integration/install_add_pod/after/Pods/Pods-header.h
@@ -1,5 +1,0 @@
-// WARNING: This feature of CocoaPods is present for discussion purposes and might be discontinued or changed in future
-#define __COCOA_PODS
-
-#define __POD_JSONKit
-#define __POD_Reachability