+++ spec/integration/install_subspecs/after/Pods/Pods-iOS App-environment.h
@@ -1,0 +1,14 @@
+
+// To check if a library is compiled with CocoaPods you
+// can use the `COCOAPODS` macro definition which is
+// defined in the xcconfigs so it is available in
+// headers also when they are imported in the client
+// project.
+
+
+// PodTest/subspec_1
+#define COCOAPODS_POD_AVAILABLE_PodTest_subspec_1 TRUE
+#define COCOAPODS_VERSION_MAJOR_PodTest_subspec_1 1
+#define COCOAPODS_VERSION_MINOR_PodTest_subspec_1 0
+#define COCOAPODS_VERSION_PATCH_PodTest_subspec_1 0
+