+++ spec/unit/library_spec.rb
@@ -45,7 +45,7 @@ module Pod
       end
 
       it "returns the absolute path of the target header file" do
-        @lib.target_header_path.to_s.should.include?('Pods/Pods-header.h')
+        @lib.target_environment_header_path.to_s.should.include?('Pods/Pods-environment.h')
       end
 
       it "returns the absolute path of the prefix header file" do